# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/connections/follow/{id}": {
      "post": {
        "operationId": "ConnectionsController_follow[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Connections"
        ]
      }
    },
    "/connections/{id}": {
      "post": {
        "operationId": "ConnectionsController_follow[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Connections"
        ]
      },
      "delete": {
        "operationId": "ConnectionsController_unfollow[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connections"
        ]
      }
    },
    "/connections/unfollow/{id}": {
      "delete": {
        "operationId": "ConnectionsController_unfollow[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connections"
        ]
      }
    },
    "/connections/followers/{id}": {
      "get": {
        "operationId": "ConnectionsController_followers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connections"
        ]
      }
    },
    "/connections/following/{id}": {
      "get": {
        "operationId": "ConnectionsController_following[0]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connections"
        ]
      }
    },
    "/connections/followeds/{id}": {
      "get": {
        "operationId": "ConnectionsController_following[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connections"
        ]
      }
    },
    "/connections/summary/{id}": {
      "get": {
        "operationId": "ConnectionsController_summary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connections"
        ]
      }
    },
...
    "/events": {
      "get": {
        "operationId": "EventsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      },
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      },
      "put": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      },
      "delete": {
        "operationId": "EventsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/repost": {
      "post": {
        "operationId": "EventsController_repost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/attendees": {
      "get": {
        "operationId": "EventsController_listAttendees",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/attend": {
      "get": {
        "operationId": "EventsController_myAttendance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      },
      "post": {
        "operationId": "EventsController_attend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      },
      "delete": {
        "operationId": "EventsController_cancelAttend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/adoptable-animals": {
      "post": {
        "operationId": "EventsController_registerAdoptableAnimal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAdoptableAnimalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/adoptable-animals/{animalId}/transfer": {
      "post": {
        "operationId": "EventsController_transferAdoptionTutorship",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "animalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferAdoptionTutorshipDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
...
  },
  "info": {
    "title": "PataNet API",
    "description": "API do ecossistema PataNet (usuarios, pets, adocao, eventos, feed e moderacao)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEventDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {}
      },
      "RegisterAdoptableAnimalDto": {
        "type": "object",
        "properties": {}
      },
      "TransferAdoptionTutorshipDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}